feat: Restructure app into organized component folders
