+++ package.json
@@ -5,11 +5,14 @@
   "dependencies": {
     "@emotion/react": "^11.11.4",
     "@emotion/styled": "^11.11.0",
+    "@fontsource/lato": "^5.2.5",
     "@mui/icons-material": "^5.15.13",
     "@mui/material": "^5.15.13",
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "ajv": "^6.12.6",
+    "ajv-keywords": "^3.5.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
@@ -40,4 +43,4 @@
       "last 1 safari version"
     ]
   }
-} 
+}
