+++ src/styles/theme.js
@@ -1,0 +1,9 @@
+import { createTheme } from "@mui/material";
+
+const theme = createTheme({
+  typography: {
+    fontFamily: "Lato, sans-serif",
+  },
+});
+
+export default theme;
